Initial commit; create a child theme and define a visual effect.
A child theme is the exact same thing as a regular WordPress theme, with
the exception that its `style.css` file contains a header line that
defines a `Template` value whose key is the parent Theme's slug (folder
name). In this case, that value is `storefront`, which is how WordPress
knows this theme is Storefront's child.

To be a "child" theme means to refer to the parent theme for any
instruction not given in the child. If the child theme contains only
this one `style.css` file and nothing else, then the effect is
equivalent to using the parent theme.

In this case, the CSS code in the `style.css` is loaded *after* the code
in the parent theme, giving us a clean way to override any styling that
the parent theme performs. For our purposes, we simply do a basic CSS
image replacement to make "Out of Stock" products show an "On Backorder"
notice. The notice is pure text, no images required, because it is
created using CSS generated content with the `::after` pseudo-element.