Fixes meitar/flora-posidonia#16: Product pages, "Tags" to "Ingredients".
This commit overrides a single, small section of the default WooCommerce
template. The entire file from the WooCommerce plugin itself is copied
herein, and a single line (the line that now contains `Ingredients`) is
edited. In other words:

    cp wp-content/plugins/woocommerce/templates/single-product/meta.php \
        wp-content/themes/flora-posidonia-storefront/woocommerce/single-product/meta.php

From there, the file is merely edited to contain the new user interface.
A Text Domain has been added so that this theme can take responsibility
to offer translations of its own UI text modifications in the future.

This commit also updates the theme's own version to version 0.1.1.